pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "grip-digest"
version = "0.1.0"
description = "Group Research Intelligence Pipeline — daily paper digest for Slack"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["research", "papers", "arxiv", "slack", "digest", "llm"]
authors = [
    { name = "Your Lab Name", email = "you@example.com" }
]

dependencies = [
    "anthropic>=0.20.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# Install when you're ready to add these sources
rss = ["feedparser>=6.0.0"]
semantic = ["requests>=2.31.0"]
# Install for the Slack Events API feedback server
feedback-server = ["flask>=3.0.0"]
# Install everything
all = [
    "feedparser>=6.0.0",
    "requests>=2.31.0",
    "flask>=3.0.0",
]
# Dev tools
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
# After `pip install`, these CLI commands become available:
grip = "grip.cli:main"
grip-update-profile = "grip.cli:update_profile"

[project.urls]
Repository = "https://github.com/your-org/grip"

# ── Hatch build config ────────────────────────────────────────────────────────
[tool.hatch.build.targets.wheel]
packages = ["src/grip"]

# Include the default data files (profile template, gitkeeps) in the package
[tool.hatch.build.targets.wheel.force-include]
"src/grip/data/interest_profile.txt" = "grip/data/interest_profile.txt"

# ── Ruff linter ───────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

# ── Mypy type checker ─────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["src/grip/tests"]
addopts = "-v --tb=short"
